adds $bootstrap_ipxe_url param
@ktreese,
I'm hoping this will allow me to overwrite the bootstrap url in order to get around authentication that is causing a 401 Access Denied when curling the bootstrap.ipxe from a Razor server that has RBAC with Shiro enabled. The idea is that I would overwrite it on my end with "localhost" instead of fqdn which will then allow me to download the file. The problem then becomes the URL embedded in the ipxe file is wrong (it points to localhost instead of the fqdn of the Razor server), this will cause PXE operations to fail. My solution to this is to then sed the file as follows: `sed 's/localhost/fqdn_of_my_server/'`